+++ sae-20190506/alibabacloud_sae20190506/client.py
@@ -1077,6 +1077,8 @@ class Client(OpenApiClient):
             query['SecurityGroupId'] = request.security_group_id
         if not UtilClient.is_unset(request.sls_configs):
             query['SlsConfigs'] = request.sls_configs
+        if not UtilClient.is_unset(request.sls_log_env_tags):
+            query['SlsLogEnvTags'] = request.sls_log_env_tags
         if not UtilClient.is_unset(request.startup_probe):
             query['StartupProbe'] = request.startup_probe
         if not UtilClient.is_unset(request.termination_grace_period_seconds):
@@ -1289,6 +1291,8 @@ class Client(OpenApiClient):
             query['SecurityGroupId'] = request.security_group_id
         if not UtilClient.is_unset(request.sls_configs):
             query['SlsConfigs'] = request.sls_configs
+        if not UtilClient.is_unset(request.sls_log_env_tags):
+            query['SlsLogEnvTags'] = request.sls_log_env_tags
         if not UtilClient.is_unset(request.startup_probe):
             query['StartupProbe'] = request.startup_probe
         if not UtilClient.is_unset(request.termination_grace_period_seconds):
@@ -4741,6 +4745,8 @@ class Client(OpenApiClient):
             query['SecurityGroupId'] = request.security_group_id
         if not UtilClient.is_unset(request.sls_configs):
             query['SlsConfigs'] = request.sls_configs
+        if not UtilClient.is_unset(request.sls_log_env_tags):
+            query['SlsLogEnvTags'] = request.sls_log_env_tags
         if not UtilClient.is_unset(request.startup_probe):
             query['StartupProbe'] = request.startup_probe
         if not UtilClient.is_unset(request.termination_grace_period_seconds):
@@ -4949,6 +4955,8 @@ class Client(OpenApiClient):
             query['SecurityGroupId'] = request.security_group_id
         if not UtilClient.is_unset(request.sls_configs):
             query['SlsConfigs'] = request.sls_configs
+        if not UtilClient.is_unset(request.sls_log_env_tags):
+            query['SlsLogEnvTags'] = request.sls_log_env_tags
         if not UtilClient.is_unset(request.startup_probe):
             query['StartupProbe'] = request.startup_probe
         if not UtilClient.is_unset(request.termination_grace_period_seconds):